pyproject.toml
[project]
name = "mvf-hack-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "crawl4ai>=0.6.3",
    "fastapi[standard]>=0.115.12",
    "groq>=0.26.0",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "ruff>=0.11.13",
    "tqdm>=4.67.1",
]
